fix: standardize money fields to Int cents
SubPayout entity Double->Int, SubContractor hourlyRate Double->Int,
PayoutService Int arithmetic, DataSeeder hourlyRate in cents, DTO update.

+++ src/main/kotlin/com/strata/ftw/domain/entity/SubContractor.kt
@@ -34,7 +34,7 @@ class SubContractor(
     var yearsExperience: Int? = null,
 
     @Column(name = "hourly_rate")
-    var hourlyRate: Double? = null,
+    var hourlyRate: Int? = null, // Amount in cents (divide by 100 for display)
 
     var verified: Boolean = false,
     var licensed: Boolean = false,
